requirements.txt
python-telegram-bot==20.6
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
cryptography==41.0.5
pytz==2023.3
pygithub==1.59.1 